pyproject.toml
[build-system]
requires = [
    "pdm-backend>=2.1.0,<3.0.0",
    "setuptools>=67.0.0,<69.0.0",
    "wheel",
    "toml"
]
build-backend = "pdm.backend"

[project]
name = "imhide"
dynamic = ["version"]
description = "A webapp to generate QR codes"
authors = [
    { name = "Yann de Thé", email = "ydethe@gmail.com" },
]
dependencies = [
    "rich~=13.9",
    "qrcode[pil]~=7.4",
    "flask~=3.0",
    "waitress~=3.0",
    "Flask-WTF~=1.2",
    "flask-bootstrap~=3.3",
    "pillow~=10.4",
    "pydantic-settings~=2.5",
    "scikit-commpy>=0.8.0",
    "numpy>=2.2.6",
    "pyturbocode>=0.1.13",
    "cryptography>=46.0.5",
]
requires-python = ">=3.12"
readme = "README.md"

[project.urls]
"Bug Tracker" = "https://https://github.com//ydethe/imhide/issues"
Homepage = "https://https://github.com//ydethe/imhide"
Source = "https://https://github.com//ydethe/imhide"

[tool.pylint.format]
max-line-length = 100

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--html=htmldoc/tests/report.html",
    "--self-contained-html",
    "--junitxml=htmldoc/junit.xml",
    "--cov-config=tests/coverage.conf",
    "--cov=imhide",
    "--doctest-modules"
]
testpaths = [
    "tests/",
]

[tool.pdm.version]
source = "scm"

[tool.pdm.options]
add = ["--save-compatible"]

[tool.pdm.build]

excludes = [
    ".gitignore",
    "tests",
    "examples",
    "build",
    "dev",
    "dist"
]
editable-backend = "editables"

[dependency-groups]
dev = [
    "coverage-badge>=1.1.0",
    "black==22.3.0",
    "ipython~=7.23.1",
    "pre-commit",
    "snakeviz>=2.1.2",

]
maintain = [
    "black==22.3.0",
    "blacken-docs==1.12.1",
    "git-changelog>=1.0",
]
quality = [
    "mypy>=1.2.0",
    "ruff>=0.0.246",
]
doc = [
    "pdoc3~=0.10.0",
    "genbadge[all]>=1.1.1,<2.0.0",
    "docstr-coverage>=2.2.0",
    "pdm_template>=0.1.8",
]
test = [
    "pytest~=7.1.3",
    "pytest-cov~=3.0.0",
    "pytest-mock~=3.8.2",
    "pytest-instafail~=0.4.2",
    "pytest-picked~=0.4.6",
    "pytest-sugar~=0.9.5",
    "pytest-html~=3.1.1",

]

[tool.pdm.scripts]
site_packages = true
doc.shell = """\
    mkdir -p htmldoc/imhide
    pdoc --html --force --config latex_math=True -o htmldoc imhide
    coverage html -d htmldoc/coverage --rcfile tests/coverage.conf
    coverage xml -o htmldoc/coverage/coverage.xml --rcfile tests/coverage.conf
    docstr-coverage src/imhide -miP -sp -is -idel --skip-file-doc --badge=htmldoc/imhide/doc_badge.svg
    genbadge coverage -l -i htmldoc/coverage/coverage.xml -o htmldoc/imhide/cov_badge.svg
    fix-doc
"""
